积分列表增加删除功能

+++ rzb-ui/src/views/core/list.vue
@@ -35,13 +35,26 @@ export default {
       })
     },
     deleteById(id) {
-      deleteById(id).then(response => {
+      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteById(id).then(response => {
+          this.$message({
+            type: 'success',
+            message: response.msg
+          });
+          this.fetchDate();
+        })
+
+      }).catch(() => {
         this.$message({
-          message: response.msg,
-          type: 'success'
+          type: 'info',
+          message: '已取消删除'
         });
-        this.fetchDate();
-      })
+      });
+
     }
   }
 }
